refactor: bookmark 페이지 및 관련 컴포넌트 관심사 분리
